+++ docs/index.rst
@@ -5,7 +5,7 @@
    quickstart
 
 
-token-auth-request
-==================
+token-auth-requests
+===================
 
-`token_auth_request` implements token based authentication for HTTP clients. To find out more, :doc:`install <installation>` the package and read the :doc:`quickstart guide <quickstart>`.
+`token_auth_requests` implements token based authentication for HTTP clients. To find out more, :doc:`install <installation>` the package and read the :doc:`quickstart guide <quickstart>`.
